+++ forui/lib/forui.dart
@@ -45,5 +45,6 @@ export 'widgets/tabs.dart';
 export 'widgets/text_field.dart';
 export 'widgets/tile.dart';
 export 'widgets/time_picker.dart';
+export 'widgets/tree.dart';
 export 'widgets/time_field.dart';
 export 'widgets/tooltip.dart';